requirements.txt
aiogram~=3.26
python-dotenv~=1.2
instagrapi~=2.3
yt-dlp~=2026.2
tgcrypto
